[#229] refactor: 토론 상세페이지 목록으로 돌아가기 버튼, 타이틀 컴포넌트 리팩토링
